requirements.txt
Django==1.7.1
argparse==1.2.1
django-crispy-forms==1.4.0
wsgiref==0.1.2
